# openapi.yaml
openapi: 3.1.0
info:
  title: Ddareungi Recommendation API
  description: 서울 따릉이 최적 대여소 추천 API
  version: 1.0.0

servers:
  - url: https://ddareungi-gpt-production.up.railway.app

paths:
  /recommend:
    post:
      operationId: getBikeRecommendation
      summary: Get bike recommendations by coordinates
      description: 위도와 경도를 기반으로 가까운 따릉이 대여소를 추천합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - lat
                - lng
              properties:
                lat:
                  type: number
                  description: 사용자 위도
                  example: 37.5665
                lng:
                  type: number
                  description: 사용자 경도
                  example: 126.978
      responses:
        "200":
          description: 추천 결과
          content:
            application/json:
              schema:
                type: object
                properties:
                  stations:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        lat:
                          type: number
                        lng:
                          type: number
                        distance:
                          type: number
                        bikeCount:
                          type: integer
                        score:
                          type: number
                        reason:
                          type: string
                        mapUrl:
                          type: string

  /recommendByPlace:
    post:
      operationId: getBikeRecommendationByPlace
      summary: Get bike recommendations by place name
      description: 장소명 또는 지역명을 기반으로 가까운 따릉이 대여소를 추천합니다.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - place
              properties:
                place:
                  type: string
                  description: 검색할 장소명 또는 지역명
                  example: "마곡역"
      responses:
        "200":
          description: 장소 기반 추천 결과
          content:
            application/json:
              schema:
                type: object
                properties:
                  place:
                    type: string
                  address:
                    type: string
                  lat:
                    type: number
                  lng:
                    type: number
                  stations:
                    type: array
                    items:
                      type: object
                      properties:
                        name:
                          type: string
                        lat:
                          type: number
                        lng:
                          type: number
                        distance:
                          type: number
                        bikeCount:
                          type: integer
                        score:
                          type: number
                        reason:
                          type: string
                        mapUrl:
                          type: string